rest-api/requirements.txt
# constraints for fairseq, works with TTS as well
numba==0.56.2
numpy>=1.23.0

fastapi
gunicorn
uvicorn
TTS

ai4bharat-transliteration
aksharamukha
